slightly changed the way how GFARM_FILE_EXCLUSIVE is handled, so that the flag can be actually used for locking.
But with this cange, the GFARM_ERR_ALREADY_EXISTS condition can happen
at not only gfs_pio_create() time, but also gfs_pio_set_view_*() time.
This may not be a problem with system call hooks, but may be a problem
with gfarm native applications.
XXX - We should revisit this GFARM_FILE_EXCLUSIVE handling.